+++ src/pages/config/schema/spamlists.rs
@@ -302,6 +302,27 @@ impl Builder<Schemas, ()> {
             .list_fields(["_id"])
             .no_list_action(Action::Modify)
             .build()
+            // Domain block list
+            .new_schema("spam-block")
+            .reload_prefix("lookup")
+            .names("domain", "domains")
+            .prefix("lookup.spam-block")
+            .new_id_field()
+            .label("Domain Name")
+            .help("The domain name to be added to the blocked domains list")
+            .input_check(
+                [Transformer::Trim],
+                [Validator::Required, Validator::IsRegex],
+            )
+            .build()
+            .new_form_section()
+            .field("_id")
+            .build()
+            .list_title("Blocked domain names")
+            .list_subtitle("Manage blocked domain names")
+            .list_fields(["_id"])
+            .no_list_action(Action::Modify)
+            .build()
             // DMARC allow list
             .new_schema("spam-dmarc")
             .reload_prefix("lookup")